Changed names & structure of all categories components
- changed names of CatSlider & mainMovies and added CatOverview and MovieCard
