+++ src/Components/Contents/CatOverview.js
@@ -1,0 +1,21 @@
+import React from "react";
+
+// === R E S E R V E D ===
+// import thumbnail from "../../Assets/images/slider/40X40.gif";
+// import Toast from 'react-bootstrap/Toast';
+// import Container from 'react-bootstrap/Container';
+// import Button from 'react-bootstrap/Button';
+// import Carousel from 'react-bootstrap/Carousel';
+
+
+
+
+const CatOverview = ({categories}) => {
+    return (
+        <>
+        {categories} 
+        </>
+    );
+};
+
+export default CatOverview;
