+++ src/Components/Contents/FilmsCategories.js
@@ -1,42 +1,42 @@
 import React from 'react'
 import movieDataRequests from "../../utilities/config";
-import Main from "./mainMovies";
+import CatSlider from "./CatSlider";
 
 export default function FilmsCategories() {
     return (
         <>
-            <Main
+            <CatSlider
                 title="Comedy"
                 fetchUrl={movieDataRequests.fetchComedy}
                 isLargeRow
             />
 
-            <Main
+            <CatSlider
                 title="Action"
                 fetchUrl={movieDataRequests.fetchAction}
                 isLargeRow
             />
-             <Main
+             <CatSlider
                 title="Thriller"
                 fetchUrl={movieDataRequests.fetchThriller}
                 isLargeRow
             />
-             <Main
+             <CatSlider
                 title="Family"
                 fetchUrl={movieDataRequests.fetchFamily}
                 isLargeRow
             />
-             <Main
+             <CatSlider
                 title="Fantasy  "
                 fetchUrl={movieDataRequests.fetchFantasy}
                 isLargeRow
             />
-             <Main
+             <CatSlider
                 title="Crime"
                 fetchUrl={movieDataRequests.fetchCrime}
                 isLargeRow
             />
-               <Main
+               <CatSlider
                 title="Adventure"
                 fetchUrl={movieDataRequests.fetchAdventure}
                 isLargeRow
