+++ src/Components/Contents/MovieCard.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function MovieCard() {
+    return (
+        <div>
+            
+        </div>
+    )
+}
