+++ src/Components/Pages/MainPages/FilmsScreen.jsx
@@ -4,7 +4,7 @@ import Header from "../../../Components/Layout/Header";
 import Footer from "../../../Components/Layout/Footer";
 
 import FeaturedSlider from "../../Contents/FeaturedSlider";
-import CatSlider from "../../Contents/CatSlider";
+import CatOverview from "../../Contents/CatOverview";
 import FilmsCategories from "../../Contents/FilmsCategories";
 import FilmsImage from "../../../Assets/images/slider/netflix-films-banner-01.jpg";
 
@@ -14,7 +14,7 @@ const Films = () => {
     <Header />
     <div className="films-container">
       <FeaturedSlider image={FilmsImage} />
-      <CatSlider categories={<FilmsCategories/>}/>
+      <CatOverview categories={<FilmsCategories/>}/>
     </div>
     <Footer />
     </>
