+++ src/Components/Pages/MainPages/HomeScreen.jsx
@@ -4,7 +4,7 @@ import Header from "../../../Components/Layout/Header";
 import Footer from "../../../Components/Layout/Footer";
 
 import FeaturedSlider from "../../Contents/FeaturedSlider";
-import CatSlider from "../../Contents/CatSlider";
+import CatOverview from "../../Contents/CatOverview";
 import HomeCategories from "../../Contents/HomeCategories";
 import HomeImage from "../../../Assets/images/slider/netflix-home-banner-02.jpg";
 
@@ -14,7 +14,7 @@ const Home = (props) => {
       <Header />
       <div className="home-container">
         <FeaturedSlider image={HomeImage} />
-        <CatSlider categories={<HomeCategories/>}/>
+        <CatOverview categories={<HomeCategories/>}/>
       </div>
 
       <Footer />
